Moved the accessing of the uploadedFiles.json to its own package
